Move Consistency, SerialConsistencty, TryFromPrimitiveError to core
