+++ scylla-cql/src/frame/mod.rs
@@ -20,7 +20,6 @@ use frame_errors::{
     CqlRequestSerializationError, FrameBodyExtensionsParseError, FrameHeaderParseError,
     LowLevelDeserializationError,
 };
-use thiserror::Error;
 use tokio::io::{AsyncRead, AsyncReadExt};
 use uuid::Uuid;
 
@@ -49,7 +48,9 @@ pub mod flag {
 }
 
 // Re-export stable public types from scylla-cql-core.
-pub use scylla_cql_core::frame::{Authenticator, Compression, CompressionFromStrError};
+pub use scylla_cql_core::frame::{
+    Authenticator, Compression, CompressionFromStrError, TryFromPrimitiveError,
+};
 
 /// A serialized CQL request frame, nearly ready to be sent over the wire.
 ///
@@ -318,24 +319,6 @@ pub fn decompress(
         Compression::Snappy => snap::raw::Decoder::new()
             .decompress_vec(comp_body)
             .map_err(|err| FrameBodyExtensionsParseError::SnapDecompressError(Arc::new(err))),
-    }
-}
-
-/// An error type for parsing an enum value from a primitive.
-#[derive(Error, Debug, Clone, PartialEq, Eq)]
-#[error("No discrimant in enum `{enum_name}` matches the value `{primitive:?}`")]
-pub struct TryFromPrimitiveError<T: Copy + std::fmt::Debug> {
-    enum_name: &'static str,
-    primitive: T,
-}
-
-impl<T: Copy + std::fmt::Debug> TryFromPrimitiveError<T> {
-    /// Creates a new `TryFromPrimitiveError` with the given enum name and primitive value.
-    pub fn new(enum_name: &'static str, primitive: T) -> Self {
-        Self {
-            enum_name,
-            primitive,
-        }
     }
 }
